flask/requirements.txt
bidict==0.21.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-SocketIO==5.1.0
google-api-core==1.30.0
google-auth==1.31.0
google-cloud==0.34.0
google-cloud-speech==2.4.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
libcst==0.3.19
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.19.5
packaging==20.9
pkg-resources==0.0.0
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
SpeechRecognition==3.8.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
